Version 1.0.0 commit Fully functional but needs some cosmetic fine-tuning.
Need to implement the top/bottom/fade settings for removing and inserting rows.

Need to pull out a ton of debug printing and put in NSLog statements that properly
report progress under control of UITABLEVIEW_REORDER_DEVELOPMENT_PROJECT
define so they are not there when used in other projects by default.

Need to strip out a bunch of old ‘ifdef’d out stuff!